Rename directory og css
