+++ src/sider/journalforing/komponenter/avsender/avsender.less
@@ -1,7 +1,3 @@
-.linje {
-  margin-bottom: 0.3em;
-}
-
 .avsender {
   background: lightgray;
   border: 1px solid #b7b1a9;
@@ -13,14 +9,18 @@
     font-weight: bold;
   }
 
+  &__input {
+    margin-bottom: 1em;
+  }
+
   &__navn {
     display: flex;
+    margin-bottom: 0.3em;
   }
 }
 
-.avsendervelger__radio {
-  margin-bottom: 1em;
-}
-avsender__input {
-  margin-bottom: 1em;
+.avsendervelger {
+  &__radio {
+    margin-bottom: 1em;
+  }
 }
