+++ src/sider/journalforing/komponenter/preutfyltAvsender.js
@@ -3,7 +3,7 @@ import PT from "prop-types";
 
 import * as Nav from "../../../navFrontend";
 
-import "./avsendere/avsendervelger.css";
+import "./avsender/avsender.css";
 
 const PreutfyltAvsender = ({ className, avsenderID, avsenderNavn }) => (
   <div className={className}>
